Moves parameters to param_keys
